feat: add functionality to mark tasks as in progress or done
This commit adds the ability to mark tasks as in progress or done in the task manager CLI application. The `mark_in_progress` function updates the status of a task to "in-progress" and saves the changes to the data.json file. Similarly, the `mark_done` function updates the status of a task to "done" and also saves the changes to the data.json file.

These changes enhance the task management capabilities of the application, allowing users to track the progress of their tasks more effectively.